read time series points from view
