--- source/geh_calculated_measurements/src/geh_calculated_measurements/electrical_heating/infrastructure/measurements/__init__.py
@@ -1,5 +1,0 @@
-from .repository import Repository
-
-__all__ = [
-    "Repository",
-]
