+++ source/geh_calculated_measurements/src/geh_calculated_measurements/electrical_heating/infrastructure/measurements_gold/database_definitions.py
@@ -1,3 +1,3 @@
 class MeasurementsGoldDatabaseDefinition:
     DATABASE_NAME = "measurements_gold"
-    TIME_SERIES_POINTS_NAME = "time_series_points_v1"
+    TIME_SERIES_POINTS_NAME = "electrical_heating_v1"
